client/client/pyproject.toml
[tool.poetry]
name = "blindllamav2"
version = "0.1.3"
description = ""
authors = ["Mithril Security <contact@mithrilsecurity.io>"]
readme = "README.md"
include = ["Azure Virtual TPM Root Certificate Authority 2023.crt"]

[tool.poetry.dependencies]
python = "^3.11"
requests = { version = "^2.28.2"}
pydantic = { version = "^2.2.0" }
pyopenssl = { version = "^23.2.0" }
rich = { version = "^13.3.5" }
pyyaml = { version = "^6.0" }

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pydeps = "^1.12.17"
isort = "^5.12.0"
mypy = "^1.6.1"
types-requests = "^2.31.0.10"
types-pyyaml = "^6.0.12.12"
types-pyopenssl = "^23.3.0.0"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]
files = "blindllamav2"
check_untyped_defs = true
# ignore_missing_imports = True

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--ignore=tests/integration",
]

[tool.ruff.lint]
# Enable all `pydocstyle` rules, limiting to those that adhere to the
# Google convention via `convention = "google"`, below.
select = ["D"]

# On top of the Google convention, disable `D417`, which requires
# documentation for every function parameter.
ignore = ["D417",
 "D100",
 "D101",
 "D102",
 "D103",
 "D104",
 "D105",
 "D107",
]

[tool.ruff.lint.pydocstyle]
convention = "google"